requirements.txt
playwright
requests
python-dotenv
google-generativeai
google-genai
openai-whisper
SpeechRecognition
pyaudio
pygame
